implemented welcome db in guildMemberAdd
- refactor of sendWelcomeWith
- implemented support to imageMock

+++ src/events/guildMemberAdd/welcome/index.ts
@@ -2,7 +2,8 @@ import { type Webhook, type GuildMember } from 'discord.js'
 import { type Canvas } from '../../../commands/setWelcome/validate'
 import { readFileSync } from 'node:fs'
 import path from 'node:path'
-import buildWelcomeImage from '../../../shared/buildWelcomeImage'
+import db from '../../../db'
+import SendWelcomeWith from '../../../shared/sendWelcomeWith'
 interface Options {
   channelId: string
 }
@@ -17,9 +18,13 @@ async function buildWebhook(member: GuildMember, options: Options): Promise<Webh
   return webhook
 }
 export default async function welcome(member: GuildMember): Promise<void> {
-  const mock: Canvas = JSON.parse(readFileSync(path.join(__dirname, '../../../mocks/welcome.json'), 'utf-8'))
-  const welcomeImage = await buildWelcomeImage(mock, member)
-  const webhook = await buildWebhook(member, { channelId: '951875024363585558' })
-  await webhook?.send({ content: 'Welcome to', files: [welcomeImage] })
+  const welcomeDb = await db.welcomeCommand.findUnique({
+    where: { serverId: member.guild.id }
+  })
+  if (!welcomeDb) return
+  const { channelId, image, message, send } = welcomeDb
+  const imageMock: Canvas = JSON.parse(readFileSync(path.join(__dirname, '../../../mocks/welcome.json'), 'utf-8'))
+  const webhook = await buildWebhook(member, { channelId })
+  await webhook?.send(await SendWelcomeWith({ send, image: image ?? imageMock, message, member }))
   await webhook.delete()
 }
